backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq==0.11.0
pillow==10.4.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.9.1
